requirements.txt
# Web scraping
requests
beautifulsoup4
tqdm
python-dotenv


flask




langchain>=0.2
langchain-community>=0.2
langchain-huggingface>=0.0.3

transformers>=4.42.0
faiss-cpu




fastapi
uvicorn

